Fix small bug from curl 404 output
This PR fixes a small bug from when `curl` receives an `HTTP 404` from the web server and the error code is non-zero (in this case `22`). This handles that case in the if statement. There may other error codes that `curl` throws, HTTP-wise, that you may want to catch in the future as well but I didn't run into them.